--- Go/go_best_practice/csv_query/config.toml
@@ -1,9 +1,0 @@
-sep = ","
-
-[Head]
-path = "data/owid-covid-data.csv"
-# type [bool, int, float, string, date]
-# fields = [
-#    {name = "first_name", type = "string"},
-#   {name = "date", type = "date"}
-#   ]
