+++ Go/go_best_practice/csv_query/go.mod
@@ -6,4 +6,5 @@ require (
 	github.com/pelletier/go-toml v1.9.0
 	github.com/spf13/afero v1.6.0
 	github.com/stretchr/testify v1.4.0
+	go.uber.org/zap v1.16.0
 )
