+++ Go/go_best_practice/csv_query/task.md
@@ -11,10 +11,10 @@ column_name OP value [AND/OR column_name OP value] ...
 age > 40 and status = "sick"
 ```
 - [x] При запуске программа печатает путь до исполняемого файла и версию последнего коммита
-- [ ] Программа должна корректно обрабатывать выход по сигналу SIGINT, прерывая поиск, если он запущен
-- [ ] Программа должна получать настройки из текстового конфигурационного файла (например, в TOML формате) при старте
-- [ ] Программа должна завершать исполнение запроса, если он занимает слишком продолжительное время(значение таймаута задается в конфигурационном файле)
-- [ ] Программа должна логировать все запросы в файл  access.log, логировать все ошибки (например, остановку пользователем или прервываниями по таймауту, невалидные запросы пользователя) в  error.log
+- [x] Программа должна корректно обрабатывать выход по сигналу SIGINT, прерывая поиск, если он запущен
+- [x] Программа должна получать настройки из текстового конфигурационного файла (например, в TOML формате) при старте
+- [x] Программа должна завершать исполнение запроса, если он занимает слишком продолжительное время(значение таймаута задается в конфигурационном файле)
+- [x] Программа должна логировать все запросы в файл  access.log, логировать все ошибки (например, остановку пользователем или прервываниями по таймауту, невалидные запросы пользователя) в  error.log
 - [ ] Код должен быть покрыт тестами (test coverage хотя бы 30%)
 - [ ] Код должен быть оргранизован согласно выбранным принципам, например можно использовать 
         [proeject-layout](https://github.com/golang-standards/project-layout) для вдохновления
